requirements.txt
fastapi
uvicorn
sounddevice
soundfile
numpy
google-generativeai
python-dotenv
jinja2
python-multipart
python-magic
python-magic-bin
